collapsible items bug fix
