+++ src/components/LineItem.jsx
@@ -12,14 +12,18 @@ import {
   TextField,
   Tooltip,
   Typography,
+  Button,
+  Box,
+  Collapse
 } from '@mui/material';
 import PropTypes from 'prop-types';
 import { React, useState } from 'react';
 import poLineStatusCodes from '../static/data/poLineStatusCodes.json';
 import CommentModal from './CommentModal';
-import { InfoOutlined } from '@mui/icons-material';
+import { InfoOutlined, KeyboardArrowDown } from '@mui/icons-material';
 
 function LineItem({
+  index,
   lineItem,
   lineItemError,
   mapID,
@@ -27,129 +31,141 @@ function LineItem({
   updateLineItem,
 }) {
   const [commentModalOpen, setCommentModalOpen] = useState(false);
+  const [open, setOpen] = useState(true);
+
   const handleEventChange = (event, prop) => {
     updateLineItem(mapID, prop, event.target.value);
   };
 
+  const handleLineItemCollapse = () => {
+    setOpen(!open);
+  }
+
   return (
-    <Grid container spacing={2} mb={4}>
-      <Grid item md={4}>
-        <TextField
-          fullWidth
-          required
-          error={lineItemError?.item}
-          label="Item"
-          value={lineItem.item}
-          onChange={(event) => handleEventChange(event, 'item')}
-        />
-      </Grid>
-      <Grid item md={4}>
-        <TextField
-          fullWidth
-          label="Description"
-          value={lineItem.description}
-          onChange={(event) => handleEventChange(event, 'description')}
-        />
-      </Grid>
-      <Grid item md={4}>
-        <TextField
-          fullWidth
-          required
-          error={lineItemError?.unitOfMeasure}
-          label="Unit of Measure"
-          value={lineItem.unitOfMeasure}
-          onChange={(event) => handleEventChange(event, 'unitOfMeasure')}
-        />
-      </Grid>
-      <Grid item md={4}>
-        <TextField
-          fullWidth
-          required
-          error={lineItemError?.orderedQuantity}
-          label="Ordered quantity"
-          value={lineItem.orderedQuantity}
-          onChange={(event) => handleEventChange(event, 'orderedQuantity')}
-        />
-      </Grid>
-      <Grid item md={4}>
-        <TextField
-          fullWidth
-          required
-          error={lineItemError?.acknowledgedQuantity}
-          label="Acknowledged quantity"
-          value={lineItem.acknowledgedQuantity}
-          onChange={(event) => handleEventChange(event, 'acknowledgedQuantity')}
-        />
-      </Grid>
-      <Grid item md={4}>
-        <TextField
-          fullWidth
-          required
-          error={lineItemError?.price}
-          label="Price"
-          value={lineItem.price}
-          onChange={(event) => handleEventChange(event, 'price')}
-          InputProps={{
-            endAdornment: (
-              <InputAdornment position="end">
-                <Tooltip title="must be a number" arrow>
-                  <InfoOutlined />
-                </Tooltip>
-              </InputAdornment>
-            )
-          }}
-        />
-      </Grid>
-      <Grid item xs={4}>
-        <FormControl fullWidth>
-          <InputLabel>Acknowledgement Status</InputLabel>
-          <Select
-            label="Acknowledgement Status"
-            value={lineItem.acknowledgementStatus}
-            onChange={(event) =>
-              handleEventChange(event, 'acknowledgementStatus')
-            }
-          >
-            {poLineStatusCodes.map((poLineStatusCode) => (
-              <MenuItem
-                value={poLineStatusCode.code}
-                key={poLineStatusCode.code}
+    <Box>
+      <Button variant='text' onClick={handleLineItemCollapse} endIcon={<KeyboardArrowDown />}>{`Line Item: ${index}`}</Button>
+      <Collapse in={open} timeout="auto" unmountOnExit>
+        <Grid container spacing={2} mb={4}>
+          <Grid item md={4}>
+            <TextField
+              fullWidth
+              required
+              error={lineItemError?.item}
+              label="Item"
+              value={lineItem.item}
+              onChange={(event) => handleEventChange(event, 'item')}
+            />
+          </Grid>
+          <Grid item md={4}>
+            <TextField
+              fullWidth
+              label="Description"
+              value={lineItem.description}
+              onChange={(event) => handleEventChange(event, 'description')}
+            />
+          </Grid>
+          <Grid item md={4}>
+            <TextField
+              fullWidth
+              required
+              error={lineItemError?.unitOfMeasure}
+              label="Unit of Measure"
+              value={lineItem.unitOfMeasure}
+              onChange={(event) => handleEventChange(event, 'unitOfMeasure')}
+            />
+          </Grid>
+          <Grid item md={4}>
+            <TextField
+              fullWidth
+              required
+              error={lineItemError?.orderedQuantity}
+              label="Ordered quantity"
+              value={lineItem.orderedQuantity}
+              onChange={(event) => handleEventChange(event, 'orderedQuantity')}
+            />
+          </Grid>
+          <Grid item md={4}>
+            <TextField
+              fullWidth
+              required
+              error={lineItemError?.acknowledgedQuantity}
+              label="Acknowledged quantity"
+              value={lineItem.acknowledgedQuantity}
+              onChange={(event) => handleEventChange(event, 'acknowledgedQuantity')}
+            />
+          </Grid>
+          <Grid item md={4}>
+            <TextField
+              fullWidth
+              required
+              error={lineItemError?.price}
+              label="Price"
+              value={lineItem.price}
+              onChange={(event) => handleEventChange(event, 'price')}
+              InputProps={{
+                endAdornment: (
+                  <InputAdornment position="end">
+                    <Tooltip title="must be a number" arrow>
+                      <InfoOutlined />
+                    </Tooltip>
+                  </InputAdornment>
+                )
+              }}
+            />
+          </Grid>
+          <Grid item xs={4}>
+            <FormControl fullWidth>
+              <InputLabel>Acknowledgement Status</InputLabel>
+              <Select
+                label="Acknowledgement Status"
+                value={lineItem.acknowledgementStatus}
+                onChange={(event) =>
+                  handleEventChange(event, 'acknowledgementStatus')
+                }
               >
-                {poLineStatusCode.code +
-                  ' (' +
-                  poLineStatusCode.description +
-                  ')'}
-              </MenuItem>
-            ))}
-          </Select>
-        </FormControl>
-      </Grid>
-      <Grid item md={4}>
-        <IconButton onClick={() => setCommentModalOpen(true)}>
-          {lineItem.comment ? <CommentIcon /> : <AddCommentIcon />}
-        </IconButton>
-      </Grid>
-      <Grid item xs={12}>
-        <IconButton onClick={() => removeLineItem(mapID)}>
-          <RemoveCircleIcon sx={{ color: 'red' }} />
-          <Typography>Remove line item</Typography>
-        </IconButton>
-      </Grid>
-      {commentModalOpen && (
-        <CommentModal
-          commentModalOpen={commentModalOpen}
-          setCommentModalOpen={setCommentModalOpen}
-          comment={lineItem.comment}
-          setComment={handleEventChange}
-          header={'Line Comment - ' + lineItem.item}
-          target={'comment'}
-        />
-      )}
-    </Grid>
+                {poLineStatusCodes.map((poLineStatusCode) => (
+                  <MenuItem
+                    value={poLineStatusCode.code}
+                    key={poLineStatusCode.code}
+                  >
+                    {poLineStatusCode.code +
+                      ' (' +
+                      poLineStatusCode.description +
+                      ')'}
+                  </MenuItem>
+                ))}
+              </Select>
+            </FormControl>
+          </Grid>
+          <Grid item md={4}>
+            <IconButton onClick={() => setCommentModalOpen(true)}>
+              {lineItem.comment ? <CommentIcon /> : <AddCommentIcon />}
+            </IconButton>
+          </Grid>
+          <Grid item xs={12}>
+            <IconButton onClick={() => removeLineItem(mapID)}>
+              <RemoveCircleIcon sx={{ color: 'red' }} />
+              <Typography>Remove line item</Typography>
+            </IconButton>
+          </Grid>
+          {commentModalOpen && (
+            <CommentModal
+              commentModalOpen={commentModalOpen}
+              setCommentModalOpen={setCommentModalOpen}
+              comment={lineItem.comment}
+              setComment={handleEventChange}
+              header={'Line Comment - ' + lineItem.item}
+              target={'comment'}
+            />
+          )}
+        </Grid>
+      </Collapse>
+    </Box>
   );
 }
 
 LineItem.propTypes = {
+  index: PropTypes.number.isRequired,
   lineItem: PropTypes.shape({
     item: PropTypes.string,
     description: PropTypes.string,
