+++ src/components/LineItems.jsx
@@ -1,10 +1,8 @@
-import { KeyboardArrowDown } from '@mui/icons-material';
 import AddCircleIcon from '@mui/icons-material/AddCircle';
-import { Button, Collapse, Grid, IconButton, Typography } from '@mui/material';
+import { Grid, IconButton, Typography } from '@mui/material';
 import { ListItem, List } from '@mui/material'; //TODO MAKE SURE TO CHANGE THIS IMPORT
 import { nanoid } from 'nanoid';
 import PropTypes from 'prop-types';
-import { useState } from 'react';
 import { React } from 'react';
 import LineItem from './LineItem';
 
@@ -36,12 +34,6 @@ function LineItems({ lineItems, lineItemErrors, setLineItems }) {
     setLineItems(newLineItems);
   };
 
-  const [open, setOpen] = useState(true);
-
-  const handleLineItemCollapse = () => {
-    setOpen(!open);
-  }
-
   return (
     <Grid container>
       <Grid item xs={12}>
@@ -50,18 +42,14 @@ function LineItems({ lineItems, lineItemErrors, setLineItems }) {
         <List>
           {lineItems && Array.from(lineItems).map(([key, lineItem], index) => (
             <ListItem key={key} disableGutters>
-              <div>
-                <Button variant='text' onClick={handleLineItemCollapse} endIcon={<KeyboardArrowDown />}>{`Line Item: ${index}`}</Button>
-                <Collapse in={open} timeout="auto" unmountOnExit>
-                  <LineItem
-                  lineItem={lineItem}
-                  lineItemError={lineItemErrors.get(key)}
-                  mapID={key}
-                  removeLineItem={removeLineItem}
-                  updateLineItem={updateLineItem}
-                />
-                </Collapse>
-              </div>
+              <LineItem
+                index={index}
+                lineItem={lineItem}
+                lineItemError={lineItemErrors.get(key)}
+                mapID={key}
+                removeLineItem={removeLineItem}
+                updateLineItem={updateLineItem}
+              />
             </ListItem>
           ))}
         </List>
